requirements.txt
py-zabbix
PyYaml